# api/openapi.yaml
openapi: 3.0.3
info:
  title: FCS Task API
  version: "1.0"
  description: API for FCS Task with user management
servers:
  - url: /v1
    description: API v1
paths:
...
  /v1/users:
    post:
      operationId: CreateUser
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/signout:
    post:
      operationId: SignOut
      summary: Sign out and delete current session
      responses:
        "204":
          description: Successfully signed out
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: uuid
      description: Session token obtained from /api/signup or /api/signin
  schemas:
    User:
      $ref: "./schemas/User.yaml"
    CreateUserRequest:
      $ref: "./schemas/CreateUserRequest.yaml"
    Error:
      $ref: "./schemas/Error.yaml"
...